+++ src/test/scala/test/com/azaptree/application/ApplicationLauncherTest.scala
@@ -131,7 +131,7 @@ class TestApplicationLauncher extends ApplicationLauncher {
     group = "GROUP-2", name = "appHealthCheck3",
     config = Some(ConfigFactory.parseString("""compName = "CompC" """)))("appHealthCheck3")
 
-  val checkCompStartedScorer: (ApplicationService, HealthCheck) => (Int, Option[String]) = (appService, healthCheck) => {
+  val checkCompStartedScorer: (ApplicationService, HealthCheck) => (Short, Option[String]) = (appService, healthCheck) => {
     val compName = healthCheck.config.get.getString("compName")
     if (appService.isComponentStarted(compName)) (100, None) else (0, Some(s"$compName is not started"))
   }
